Looks like I need to wait until the rewritten app and stuff is removed? all errors point *around* there, which gives me doubts
