+++ apps/repl/app/app.ts
@@ -1,12 +1,8 @@
+import './css/tailwind.css';
+import './css/app.css';
 import 'limber-ui/theme.css';
 
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore
-import { DEBUG } from '@glimmer/env';
 import Application from '@ember/application';
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore
-import { _backburner } from '@ember/runloop';
 
 import loadInitializers from 'ember-load-initializers';
 import Resolver from 'ember-resolver';
@@ -14,16 +10,6 @@ import { setupComponentMachines } from 'ember-statechart-component';
 import { StateNode } from 'xstate';
 
 import config from 'limber/config/environment';
-
-if (DEBUG) {
-  // This has performance implications, but the debuggability is worth it.
-  // Debugging eventloop stuff is notoriously difficult, so let's make it
-  // not difficult by default.
-  //
-  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-  // @ts-ignore
-  _backburner.DEBUG = true;
-}
 
 // @babel/traverse (from babel-plugin-ember-template-imports)
 // accesses process.....
