+++ apps/repl/app/css/app.css
@@ -1,0 +1,62 @@
+* {
+  touch-action: manipulation;
+}
+
+/**
+ * https://www.w3schools.com/howto/howto_css_loader.asp
+ */
+
+.loader {
+  border: 4px solid var(--ember-faint-gray);
+  border-radius: 50%;
+  border-top: 4px solid var(--code-keyword);
+}
+
+.copy-menus-hidden [data-test-copy-menu] {
+  display: none;
+}
+
+.glimdown-render {
+  margin-top: 0.5rem;
+}
+
+.glimdown-render button {
+  color: white;
+  border-radius: 0.25rem;
+  padding: 0.25rem 0.5rem;
+  background: var(--code-bg);
+  border: 1px solid var(--horizon-border);
+}
+
+.glimdown-render button:focus {
+  outline: 2px solid transparent;
+  outline-offset: 2px;
+  --tw-ring-offset-shadow: var(--tw-ring-inset) 0 0 0 var(--tw-ring-offset-width)
+    var(--tw-ring-offset-color);
+  --tw-ring-shadow: var(--tw-ring-inset) 0 0 0 calc(4px + var(--tw-ring-offset-width))
+    var(--tw-ring-color);
+  box-shadow: var(--tw-ring-offset-shadow), var(--tw-ring-shadow), var(--tw-shadow, 0 0 #0000);
+}
+
+.glimdown-render button:focus-visible {
+  outline: 2px solid transparent;
+  outline-offset: 2px;
+}
+
+.glimdown-render button:hover {
+  opacity: 0.9;
+}
+
+.gutter {
+  background: var(--code-keyword);
+}
+.gutter-horizontal {
+  cursor: col-resize;
+}
+.gutter-vertical {
+  cursor: row-resize;
+}
+
+dialog::backdrop {
+  backdrop-filter: blur(1px);
+}
