+++ apps/repl/app/index.html
@@ -1,67 +1,64 @@
 <!doctype html>
 <html lang="en-us">
-  <head>
-    <meta charset="utf-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <title>Limber » Ember Markdown Playground</title>
-    <meta
-      name="description"
-      content="Limber is a Glimdown playground for demonstrating how Glimmer and Markdown can be used together to write documentation"
-    />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
 
-    {{content-for "head"}} {{content-for "head-footer"}}
+<head>
+  <meta charset="utf-8" />
+  <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+  <title>Limber » Ember Markdown Playground</title>
+  <meta name="description"
+    content="Limber is a Glimdown playground for demonstrating how Glimmer and Markdown can be used together to write documentation" />
+  <meta name="viewport" content="width=device-width, initial-scale=1" />
 
-    <!--
+  {{content-for "head"}} {{content-for "head-footer"}}
+
+  <!--
       This the source of truth for the ember brand variables
         https://emberjs.com/images/brand/Ember-Brand-Guidelines.pdf
     -->
-    <style type="text/css">
-      :root {
-        --ember-brand: #e04e39;
-        --ember-black: #212121;
-        --ember-burnt-ember: #9b2918;
-        --ember-gray: #817f7f;
-        --ember-blue: #1e719b;
-        --ember-faint-gray: #efebea;
-        --ember-light-blue: #74b0ce;
-        --ember-linen: #fdf7f6;
-        --ember-yellow: #fbc840;
-        --ember-white: #fdfdfd;
+  <style type="text/css">
+    :root {
+      --ember-brand: #e04e39;
+      --ember-black: #212121;
+      --ember-burnt-ember: #9b2918;
+      --ember-gray: #817f7f;
+      --ember-blue: #1e719b;
+      --ember-faint-gray: #efebea;
+      --ember-light-blue: #74b0ce;
+      --ember-linen: #fdf7f6;
+      --ember-yellow: #fbc840;
+      --ember-white: #fdfdfd;
+    }
+  </style>
+  <script>
+    globalThis.v8 = {};
+    globalThis.path = {
+      resolve: (...args) => args.join("/"),
+    };
+  </script>
+</head>
+
+<body class="overflow-hidden syntax-dark syntax-bright">
+  <div id="initial-loader">
+    <style>
+      body,
+      html {
+        margin: 0;
+        padding: 0;
       }
     </style>
-    <script>
-      globalThis.v8 = {};
-      globalThis.path = {
-        resolve: (...args) => args.join("/"),
-      };
-    </script>
-  </head>
-  <body class="overflow-hidden syntax-dark syntax-bright">
-    <div id="initial-loader">
-      <style>
-        body,
-        html {
-          margin: 0;
-          padding: 0;
-        }
-      </style>
-      <!--
+    <!--
         This is the App Shell
         It's sole purpose is to show something when the HTML loads (which is small and loads / is parsed fast)
         This gives the user an illusion of speed
       -->
-      <div style="height: 100vh; width: 100vw; display: grid; grid-template-rows: min-content 1fr">
-        <!-- header -->
-        <div style="background: var(--ember-black); height: 45px; width: 100%"></div>
+    <div style="height: 100vh; width: 100vw; display: grid; grid-template-rows: min-content 1fr">
+      <!-- header -->
+      <div style="background: var(--ember-black); height: 45px; width: 100%"></div>
 
-        <!-- everything else -->
-        <div
-          style="display: grid; grid-template-columns: auto 1fr; font-weight: bold; font-size: 24px"
-        >
-          <!-- editor -->
-          <div
-            style="
+      <!-- everything else -->
+      <div style="display: grid; grid-template-columns: auto 1fr; font-weight: bold; font-size: 24px">
+        <!-- editor -->
+        <div style="
               width: 40vw;
               background: rgba(39, 39, 42);
               color: white;
@@ -69,44 +66,41 @@
               display: grid;
               align-items: center;
               justify-content: end;
-            "
-          >
-            Loading
-          </div>
+            ">
+          Loading
+        </div>
 
-          <!-- preview -->
-          <div
-            style="
+        <!-- preview -->
+        <div style="
               padding-left: 2px;
               color: rgb(39, 39, 42);
               display: grid;
               align-items: center;
               justify-content: start;
-            "
-          >
-            Limber
-          </div>
+            ">
+          Limber
         </div>
       </div>
     </div>
+  </div>
 
-    {{content-for "body"}}
+  {{content-for "body"}}
 
-    <script>
-      if (window.self !== window.parent) {
-        window.self.document.querySelector("#initial-loader").style.display = "none";
-        window.self.document.body.classList.remove("overflow-hidden");
-      }
-    </script>
+  <script>
+    if (window.self !== window.parent) {
+      window.self.document.querySelector("#initial-loader").style.display = "none";
+      window.self.document.body.classList.remove("overflow-hidden");
+    }
+  </script>
 
-    <link integrity="" rel="stylesheet" href="{{rootURL}}assets/tailwind.css" />
-    <link integrity="" rel="stylesheet" href="{{rootURL}}assets/vendor.css" />
-    <link integrity="" rel="stylesheet" href="{{rootURL}}assets/limber.css" />
-    <script src="{{rootURL}}assets/vendor.js"></script>
-    <script src="{{rootURL}}assets/limber.js"></script>
+  <link integrity="" rel="stylesheet" href="{{rootURL}}assets/vendor.css" />
+  <link integrity="" rel="stylesheet" href="{{rootURL}}assets/limber.css" />
+  <script src="{{rootURL}}assets/vendor.js"></script>
+  <script src="{{rootURL}}assets/limber.js"></script>
 
-    {{content-for "body-footer"}}
+  {{content-for "body-footer"}}
 
-    <a rel="me" style="display: none" href="https://mastodon.coffee/@nullvoxpopuli">Mastodon</a>
-  </body>
+  <a rel="me" style="display: none" href="https://mastodon.coffee/@nullvoxpopuli">Mastodon</a>
+</body>
+
 </html>
