+++ apps/repl/package.json
@@ -67,8 +67,8 @@
     "@fortawesome/free-solid-svg-icons": "^6.5.1",
     "@glint/core": "1.3.0",
     "@glint/environment-ember-loose": "1.3.0",
-    "@glint/environment-ember-template-imports": "1.3.0",
-    "@glint/template": "1.3.0",
+    "@glint/environment-ember-template-imports": "1.4.0",
+    "@glint/template": "1.4.0",
     "@nullvoxpopuli/eslint-configs": "^3.2.2",
     "@nullvoxpopuli/limber-untyped": "workspace:^0.0.1",
     "@rollup/plugin-babel": "^6.0.4",
@@ -84,8 +84,9 @@
     "@types/qunit": "^2.19.10",
     "@types/rsvp": "^4.0.9",
     "@types/unist": "^3.0.2",
-    "@typescript-eslint/eslint-plugin": "^6.20.0",
+    "@typescript-eslint/eslint-plugin": "^7.3.1",
     "@typescript-eslint/parser": "^6.20.0",
+    "autoprefixer": "^10.4.17",
     "broccoli-asset-rev": "^3.0.0",
     "broccoli-merge-trees": "^4.2.0",
     "build-time-reporter-webpack-plugin": "^1.4.3",
@@ -121,6 +122,7 @@
     "loader.js": "^4.7.0",
     "onp": "^2.0.4",
     "pnpm-sync-dependencies-meta-injected": "^0.0.10",
+    "postcss": "^8.4.32",
     "prettier": "^3.2.5",
     "prettier-plugin-ember-template-tag": "2.0.0",
     "process": "^0.11.10",
@@ -129,12 +131,13 @@
     "qunit-console-grouper": "^0.3.0",
     "qunit-dom": "3.0.0",
     "rollup": "^4.9.6",
+    "tailwindcss": "^3.3.6",
     "testem-failure-only-reporter": "^1.0.0",
     "type-fest": "4.13.0",
     "typescript": "^5.3.3",
     "vfile": "^6.0.1",
     "vite": "^5.0.12",
-    "webpack": "5.90.3",
+    "webpack": "5.91.0",
     "yn": "^5.0.0"
   },
   "engines": {
