+++ apps/repl/tailwind.config.mjs
@@ -1,0 +1,3 @@
+import { config } from "@nullvoxpopuli/limber-styles";
+
+export default config(import.meta);
