+++ apps/repl/tests/index.html
@@ -1,37 +1,39 @@
 <!doctype html>
 <html>
-  <head>
-    <meta charset="utf-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <title>Limber Tests</title>
-    <meta name="description" content="" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
 
-    {{content-for "head"}} {{content-for "test-head"}}
+<head>
+  <meta charset="utf-8" />
+  <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+  <title>Limber Tests</title>
+  <meta name="description" content="" />
+  <meta name="viewport" content="width=device-width, initial-scale=1" />
 
-    <link integrity="" rel="stylesheet" href="{{rootURL}}assets/tailwind.css" />
-    <link rel="stylesheet" href="{{rootURL}}assets/vendor.css" />
-    <link rel="stylesheet" href="{{rootURL}}assets/limber.css" />
-    <link rel="stylesheet" href="{{rootURL}}assets/test-support.css" />
+  {{content-for "head"}} {{content-for "test-head"}}
 
-    {{content-for "head-footer"}} {{content-for "test-head-footer"}}
-  </head>
-  <body class="overflow-hidden syntax-dark syntax-bright">
-    {{content-for "body"}} {{content-for "test-body"}}
+  <link rel="stylesheet" href="{{rootURL}}assets/vendor.css" />
+  <link rel="stylesheet" href="{{rootURL}}assets/limber.css" />
+  <link rel="stylesheet" href="{{rootURL}}assets/test-support.css" />
 
-    <div id="qunit"></div>
-    <div id="qunit-fixture">
-      <div id="ember-testing-container">
-        <div id="ember-testing"></div>
-      </div>
+  {{content-for "head-footer"}} {{content-for "test-head-footer"}}
+</head>
+
+<body class="overflow-hidden syntax-dark syntax-bright">
+  {{content-for "body"}} {{content-for "test-body"}}
+
+  <div id="qunit"></div>
+  <div id="qunit-fixture">
+    <div id="ember-testing-container">
+      <div id="ember-testing"></div>
     </div>
+  </div>
 
-    <script src="/testem.js" integrity=""></script>
-    <script src="{{rootURL}}assets/vendor.js"></script>
-    <script src="{{rootURL}}assets/test-support.js"></script>
-    <script src="{{rootURL}}assets/limber.js"></script>
-    <script src="{{rootURL}}assets/tests.js"></script>
+  <script src="/testem.js" integrity=""></script>
+  <script src="{{rootURL}}assets/vendor.js"></script>
+  <script src="{{rootURL}}assets/test-support.js"></script>
+  <script src="{{rootURL}}assets/limber.js"></script>
+  <script src="{{rootURL}}assets/tests.js"></script>
 
-    {{content-for "body-footer"}} {{content-for "test-body-footer"}}
-  </body>
+  {{content-for "body-footer"}} {{content-for "test-body-footer"}}
+</body>
+
 </html>
