+++ apps/tutorial/package.json
@@ -65,14 +65,14 @@
     "@fortawesome/free-solid-svg-icons": "^6.5.1",
     "@glint/core": "1.3.0",
     "@glint/environment-ember-loose": "1.3.0",
-    "@glint/environment-ember-template-imports": "1.3.0",
-    "@glint/template": "1.3.0",
+    "@glint/environment-ember-template-imports": "1.4.0",
+    "@glint/template": "1.4.0",
     "@nullvoxpopuli/eslint-configs": "^3.2.2",
     "@nullvoxpopuli/limber-styles": "workspace:^0.0.1",
     "@nullvoxpopuli/limber-untyped": "workspace:^0.0.1",
     "@tsconfig/ember": "^3.0.2",
     "@types/qunit": "^2.19.10",
-    "@typescript-eslint/eslint-plugin": "^6.20.0",
+    "@typescript-eslint/eslint-plugin": "^7.3.1",
     "@typescript-eslint/parser": "^6.20.0",
     "broccoli-asset-rev": "^3.0.0",
     "broccoli-funnel": "^3.0.8",
@@ -109,7 +109,7 @@
     "qunit-dom": "3.0.0",
     "typescript": "^5.3.3",
     "unplugin": "^1.5.1",
-    "webpack": "5.90.3"
+    "webpack": "5.91.0"
   },
   "engines": {
     "node": ">= v16",
