+++ packages/app-support/codemirror/package.json
@@ -58,7 +58,7 @@
   "devDependencies": {
     "@babel/core": "^7.23.9",
     "@nullvoxpopuli/eslint-configs": "^3.2.2",
-    "@typescript-eslint/eslint-plugin": "^6.20.0",
+    "@typescript-eslint/eslint-plugin": "^7.3.1",
     "@typescript-eslint/parser": "^6.20.0",
     "concurrently": "^8.2.2",
     "esbuild": "0.20.2",
