+++ packages/app-support/limber-ui/addon/package.json
@@ -67,13 +67,13 @@
     "@glimmer/tracking": "^1.1.2",
     "@glint/core": "1.3.0",
     "@glint/environment-ember-loose": "1.3.0",
-    "@glint/environment-ember-template-imports": "1.3.0",
-    "@glint/template": "1.3.0",
+    "@glint/environment-ember-template-imports": "1.4.0",
+    "@glint/template": "1.4.0",
     "@nullvoxpopuli/eslint-configs": "^3.2.2",
     "@nullvoxpopuli/limber-untyped": "workspace:*",
     "@rollup/plugin-babel": "^6.0.4",
     "@tsconfig/ember": "^3.0.2",
-    "@typescript-eslint/eslint-plugin": "^6.20.0",
+    "@typescript-eslint/eslint-plugin": "^7.3.1",
     "@typescript-eslint/parser": "^6.20.0",
     "babel-plugin-ember-template-compilation": "^2.2.1",
     "concurrently": "^8.2.2",
