+++ packages/app-support/limber-ui/test-app/package.json
@@ -23,7 +23,7 @@
   },
   "dependencies": {
     "@ember/test-waiters": "^3.1.0",
-    "@glint/template": "1.3.0",
+    "@glint/template": "1.4.0",
     "ember-async-data": "1.0.3",
     "ember-modifier": "^4.1.0",
     "ember-primitives": "^0.11.4",
@@ -79,7 +79,7 @@
     "qunit": "^2.20.0",
     "qunit-dom": "3.0.0",
     "typescript": "^5.3.3",
-    "webpack": "5.90.3"
+    "webpack": "5.91.0"
   },
   "engines": {
     "node": ">= v16",
