--- packages/app-support/styles/.eslintignore
@@ -1,1 +1,0 @@
-../../../.eslintignore
