--- packages/app-support/styles/.eslintrc.cjs
@@ -1,10 +1,0 @@
-'use strict';
-
-const { configs } = require('@nullvoxpopuli/eslint-configs');
-
-const config = configs.node();
-
-module.exports = {
-  ...config,
-  overrides: [...config.overrides],
-};
