--- packages/app-support/styles/.prettierignore
@@ -1,1 +1,0 @@
-../../../.prettierignore
