--- packages/app-support/styles/.prettierrc.cjs
@@ -1,1 +1,0 @@
-../../../.prettierrc.cjs
