--- packages/app-support/styles/CHANGELOG.md
@@ -1,3 +1,0 @@
-# @nullvoxpopuli/limber-styles
-
-## 0.0.1
