--- packages/app-support/styles/broccoli-funnel.js
@@ -1,17 +1,0 @@
-'use strict';
-
-const path = require('path');
-const Funnel = require('broccoli-funnel');
-
-const SRC_FILES = path.join(__dirname, 'dist');
-
-/**
- * This broccoli funnel is for copying the built assets to a target
- * app's public folder. No building occurs
- *
- */
-module.exports = function stylesFunnel() {
-  return new Funnel(SRC_FILES, {
-    destDir: 'assets/',
-  });
-};
