--- packages/app-support/styles/turbo.json
@@ -1,8 +1,0 @@
-{
-  "extends": ["//"],
-  "pipeline": {
-    "build": {
-      "cache": false
-    }
-  }
-}
