+++ packages/app-support/transpilation/package.json
@@ -47,7 +47,7 @@
     "@types/babel__core": "^7.20.5",
     "@types/babel__standalone": "^7.1.7",
     "@types/htmlbars-inline-precompile": "^3.0.3",
-    "@typescript-eslint/eslint-plugin": "^6.20.0",
+    "@typescript-eslint/eslint-plugin": "^7.3.1",
     "@typescript-eslint/parser": "^6.20.0",
     "babel-loader": "^9.1.3",
     "ember-source": ">= 5.5.0",
@@ -56,7 +56,7 @@
     "prettier": "^3.2.5",
     "prettier-plugin-ember-template-tag": "2.0.0",
     "typescript": "^5.3.3",
-    "webpack": "5.90.3",
+    "webpack": "5.91.0",
     "webpack-cli": "^5.1.4",
     "webpack-node-externals": "^3.0.0"
   },
