+++ packages/ember-repl/addon/package.json
@@ -117,8 +117,8 @@
     "@glimmer/util": "^0.89.0",
     "@glint/core": "1.3.0",
     "@glint/environment-ember-loose": "1.3.0",
-    "@glint/environment-ember-template-imports": "1.3.0",
-    "@glint/template": "1.3.0",
+    "@glint/environment-ember-template-imports": "1.4.0",
+    "@glint/template": "1.4.0",
     "@nullvoxpopuli/eslint-configs": "^3.2.2",
     "@nullvoxpopuli/limber-untyped": "workspace:*",
     "@rollup/plugin-babel": "^6.0.4",
@@ -131,7 +131,7 @@
     "@types/mdast": "^4.0.3",
     "@types/unist": "^3.0.2",
     "@types/uuid": "^9.0.8",
-    "@typescript-eslint/eslint-plugin": "^6.20.0",
+    "@typescript-eslint/eslint-plugin": "^7.3.1",
     "@typescript-eslint/parser": "^6.20.0",
     "concurrently": "^8.2.2",
     "ember-resources": "^7.0.0",
@@ -160,6 +160,9 @@
     "version": 2,
     "type": "addon",
     "main": "addon-main.cjs",
+    "external": [
+      "@glimmer/compiler"
+    ],
     "app-js": {}
   },
   "peerDependencies": {
