+++ packages/ember-repl/test-app/package.json
@@ -48,15 +48,15 @@
     "@glimmer/tracking": "^1.1.2",
     "@glint/core": "1.3.0",
     "@glint/environment-ember-loose": "1.3.0",
-    "@glint/environment-ember-template-imports": "1.3.0",
-    "@glint/template": "1.3.0",
+    "@glint/environment-ember-template-imports": "1.4.0",
+    "@glint/template": "1.4.0",
     "@nullvoxpopuli/eslint-configs": "^3.2.2",
     "@nullvoxpopuli/limber-untyped": "workspace:^0.0.1",
     "@tsconfig/ember": "^3.0.2",
     "@types/common-tags": "^1.8.4",
     "@types/qunit": "^2.19.10",
     "@types/rsvp": "^4.0.9",
-    "@typescript-eslint/eslint-plugin": "^6.20.0",
+    "@typescript-eslint/eslint-plugin": "^7.3.1",
     "@typescript-eslint/parser": "^6.20.0",
     "concurrently": "^8.2.2",
     "ember-auto-import": "^2.7.0",
@@ -91,7 +91,7 @@
     "qunit-dom": "3.0.0",
     "typescript": "^5.3.3",
     "unified": "^10.1.2",
-    "webpack": "5.90.3"
+    "webpack": "5.91.0"
   },
   "engines": {
     "node": ">= v16",
