+++ packages/syntax/glimdown/codemirror/dev/package.json
@@ -20,7 +20,7 @@
     "@codemirror/view": "^6.23.0",
     "@glimdown/codemirror-dev-preview": "workspace:*",
     "@nullvoxpopuli/eslint-configs": "^3.2.2",
-    "@typescript-eslint/eslint-plugin": "^6.20.0",
+    "@typescript-eslint/eslint-plugin": "^7.3.1",
     "@typescript-eslint/parser": "^6.20.0",
     "eslint": "^8.55.0",
     "prettier": "^3.2.5",
