+++ packages/syntax/glimdown/codemirror/tests/package.json
@@ -8,7 +8,7 @@
     "@babel/core": "^7.23.9",
     "@babel/eslint-parser": "^7.23.10",
     "@nullvoxpopuli/eslint-configs": "^3.2.2",
-    "@typescript-eslint/eslint-plugin": "^6.20.0",
+    "@typescript-eslint/eslint-plugin": "^7.3.1",
     "@typescript-eslint/parser": "^6.20.0",
     "eslint": "^8.45.0",
     "prettier": "^3.2.5",
