+++ packages/syntax/glimmer-s-expression/lezer/package.json
@@ -25,7 +25,7 @@
     "@lezer/generator": "^1.6.0",
     "@nullvoxpopuli/eslint-configs": "^3.2.2",
     "@rollup/plugin-node-resolve": "^15.2.3",
-    "@typescript-eslint/eslint-plugin": "^6.20.0",
+    "@typescript-eslint/eslint-plugin": "^7.3.1",
     "@typescript-eslint/parser": "^6.20.0",
     "eslint": "^8.55.0",
     "prettier": "^3.2.5",
