+++ packages/syntax/glimmer/codemirror/package.json
@@ -25,7 +25,7 @@
     "@lezer/generator": "^1.6.0",
     "@nullvoxpopuli/eslint-configs": "^3.2.2",
     "@tsconfig/ember": "^3.0.2",
-    "@typescript-eslint/eslint-plugin": "^6.20.0",
+    "@typescript-eslint/eslint-plugin": "^7.3.1",
     "@typescript-eslint/parser": "^6.20.0",
     "codemirror": "^6.0.1",
     "concurrently": "^8.2.2",
